backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "src"
version = "0.1.0"
description = "IsoScape API for generating isometric city images"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.120.0,<0.130.0",
    "uvicorn==0.24.0",
    "google-genai==1.52.0",
    "python-dotenv==1.0.0",
    "pydantic>=2.12.5,<3.0.0",
    "httpx==0.28.1",
    "httpcore==1.0.9",
    "websockets==15.0.1",
    "tenacity==9.1.2",
    "typing-inspection==0.4.2"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
